requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
sentry-sdk==2.16.0
prometheus-client==0.21.0
python-dotenv==1.0.1